fix: change risk_score from int to float in SubmissionResponse
The risk_score extracted from triage_result JSON is a float (e.g. 0.35)
but the Pydantic model had risk_score: int, causing validation errors
and 500 on GET /api/v1/submissions.

+++ src/openinsure/api/submissions.py
@@ -224,7 +224,7 @@ class SubmissionResponse(BaseModel):
     updated_at: str = ""
     received_date: str = ""
     company_name: str = ""
-    risk_score: int = 0
+    risk_score: float = 0
     priority: str = "medium"
     assigned_to: str | None = None
     decision_history: list[dict[str, Any]] = Field(default_factory=list)
